+++ src/icemodel/IceModel.hh
@@ -79,6 +79,9 @@ class FloatKill;
 class HayhurstCalving;
 class CalvingAtThickness;
 class IcebergRemover;
+class CliffCalvingShear;
+class CliffCalvingTensile;
+class LinearCalving;
 }
 
 class FractureDensity;
@@ -268,6 +271,9 @@ protected:
   std::shared_ptr<calving::EigenCalving>       m_eigen_calving;
   std::shared_ptr<calving::HayhurstCalving>    m_hayhurst_calving;
   std::shared_ptr<calving::vonMisesCalving>    m_vonmises_calving;
+  std::shared_ptr<calving::CliffCalvingShear>   m_cliff_calving_shear;
+  std::shared_ptr<calving::CliffCalvingTensile> m_cliff_calving_tensile;
+  std::shared_ptr<calving::LinearCalving>       m_linear_calving;
   std::shared_ptr<PrescribedRetreat>           m_prescribed_retreat;
 
   // scalar time-dependent scaling for retreat rates coming from eigen calving, von Mises